+++ src/utils/mappers.js
@@ -15,7 +15,8 @@ import {
 import {
   CLASSIFICATION_HEAD_TITLES,
   ELECTRONIC_ACCESS_HEAD_TITLES,
-  SUBJECT_HEAD_TITLES
+  SUBJECT_HEAD_TITLES,
+  PUBLICATION_HEAD_TITLES
 } from '../components/PermissionsModal/constants/lists';
 
 
@@ -51,6 +52,8 @@ const formatData = ({ capability, column, data }) => {
       return <EmbeddedTable value={data} headTitles={SUBJECT_HEAD_TITLES} />;
     case [CAPABILITIES.INSTANCE].includes(capability) && field === CUSTOM_ENTITY_COLUMNS.CLASSIFICATION:
       return <EmbeddedTable value={data} headTitles={CLASSIFICATION_HEAD_TITLES} />;
+    case [CAPABILITIES.INSTANCE].includes(capability) && field === CUSTOM_ENTITY_COLUMNS.PUBLICATION:
+      return <EmbeddedTable value={data} headTitles={PUBLICATION_HEAD_TITLES} />;
     default:
       return data;
   }